limit processor is now fully responsible for tracking of the boundary crossings
- glyph_range no longer has width - only text_line's width is maintained;
- glyph_index_t is now unsigned int.

+++ agge.text/types.h
@@ -5,7 +5,7 @@
 
 namespace agge
 {
-	typedef uint16_t glyph_index_t;
+	typedef unsigned int glyph_index_t;
 	typedef unsigned int codepoint_t;
 
 	enum text_alignment {
