+++ samples/text-gdi/text-gdi.cpp
@@ -60,7 +60,7 @@ namespace demo
 
 						dg += j->offset;
 						for (positioned_glyphs_container_t::const_iterator k = j->begin(); k != j->end(); ++k)
-							_glyph_indices.push_back(k->index);
+							_glyph_indices.push_back(static_cast<unsigned short>(k->index));
 						::ExtTextOut(ctx, static_cast<int>(dg.dx), static_cast<int>(dg.dy), ETO_GLYPH_INDEX /*| ETO_PDY*/, 0,
 							reinterpret_cast<LPCTSTR>(&_glyph_indices[0]), static_cast<UINT>(_glyph_indices.size()), 0);
 					}
